+++ code/functions.hpp
@@ -1,0 +1,13 @@
+#ifndef FUNCTIONS_HPP
+#define FUNCTIONS_HPP
+
+
+void get_coefs(double& a, double& b, double& c);
+void find_roots(double a, double b, double c);
+
+void get_error();
+
+void print(double root1, double root2);
+void print(double root1);
+
+#endif
